Update renamed, modified, moved modified:   SubSearch/main.py renamed:    SubSearch/src/gui/assets/buttons/exit.png -> SubSearch/src/assets/buttons/exit.png renamed:    SubSearch/src/gui/assets/buttons/exit_grey.png -> SubSearch/src/assets/buttons/exit_grey.png renamed:    SubSearch/src/gui/assets/buttons/maximize.png -> SubSearch/src/assets/buttons/maximize.png renamed:    SubSearch/src/gui/assets/buttons/maximize_grey.png -> SubSearch/src/assets/buttons/maximize_grey.png renamed:    SubSearch/src/gui/assets/buttons/maximize_inactive.png -> SubSearch/src/assets/buttons/maximize_inactive.png renamed:    SubSearch/src/gui/assets/buttons/tab.png -> SubSearch/src/assets/buttons/tab.png renamed:    SubSearch/src/gui/assets/buttons/tab_grey.png -> SubSearch/src/assets/buttons/tab_grey.png renamed:    SubSearch/src/gui/assets/icons/128.ico -> SubSearch/src/assets/icons/128.ico renamed:    SubSearch/src/gui/assets/icons/16.ico -> SubSearch/src/assets/icons/16.ico renamed:    SubSearch/src/gui/assets/icons/256.ico -> SubSearch/src/assets/icons/256.ico renamed:    SubSearch/src/gui/assets/icons/32.ico -> SubSearch/src/assets/icons/32.ico renamed:    SubSearch/src/gui/assets/icons/64.ico -> SubSearch/src/assets/icons/64.ico modified:   SubSearch/src/gui/tools.py modified:   SubSearch/src/gui/widget_download.py modified:   SubSearch/src/gui/widget_root.py modified:   SubSearch/src/gui/widget_settings.py modified:   SubSearch/src/scraper/opensubtitles.py modified:   SubSearch/src/scraper/opensubtitles_soup.py modified:   SubSearch/src/scraper/subscene.py modified:   SubSearch/src/subsearch.py renamed:    SubSearch/src/__init__.py -> SubSearch/src/util/__init__.py renamed:    SubSearch/src/utilities/current_user.py -> SubSearch/src/util/current_user.py renamed:    SubSearch/src/utilities/file_manager.py -> SubSearch/src/util/file_manager.py renamed:    SubSearch/src/utilities/file_parser.py -> SubSearch/src/util/file_parser.py renamed:    SubSearch/src/utilities/local_paths.py -> SubSearch/src/util/local_paths.py renamed:    SubSearch/src/utilities/log.py -> SubSearch/src/util/log.py renamed:    SubSearch/src/utilities/raw_config.py -> SubSearch/src/util/raw_config.py renamed:    SubSearch/src/utilities/raw_registry.py -> SubSearch/src/util/raw_registry.py renamed:    SubSearch/src/utilities/string_parser.py -> SubSearch/src/util/string_parser.py renamed:    SubSearch/src/utilities/updates.py -> SubSearch/src/util/updates.py renamed:    SubSearch/src/utilities/version.py -> SubSearch/src/util/version.py deleted:    SubSearch/src/utilities/__init__.py
