+++ SubSearch/src/gui/tools.py
@@ -2,7 +2,7 @@ import ctypes
 import tkinter as tk
 
 from src.gui import tkinter_data as tkd
-from src.utilities import local_paths
+from src.util import local_paths
 
 
 GWL_EXSTYLE = -20
@@ -315,9 +315,7 @@ class ToolTip(tk.Toplevel):
         # offset the frame 1px from edge of the tooltip corner
         frame.place(x=1, y=1, anchor="nw")
         label.place(x=0, y=0, anchor="nw")
-        root_x = (
-            self.widget.winfo_rootx() + self.widget.winfo_width() + 4
-        )  # offset the tooltip by extra 4px so it doesn't overlap the widget
+        root_x = self.widget.winfo_rootx() + self.widget.winfo_width() + 4  # offset the tooltip by extra 4px so it doesn't overlap the widget
         root_y = self.widget.winfo_rooty() - self.widget.winfo_height() - 4
         # set position of the tooltip, size and add 2px around the tooltip for a 1px border
         self.geometry(f"{x+2}x{y+2}+{root_x}+{root_y}")
