+++ SubSearch/src/scraper/opensubtitles_soup.py
@@ -1,6 +1,6 @@
 import requests
 from bs4 import BeautifulSoup
-from src.utilities import log
+from src.util import log
 
 
 # search for file hash
@@ -18,29 +18,15 @@ def search_for_hash(url: str, lang: str, hi: str):
         if lang in tl:
             hi = item.find("img", alt="Subtitles for hearing impaired")
             if hi is not None and hi == "False":
-                log.output(
-                    f"Found HI-subtitle but skipping, 'cus hearing impaired is set to '{hi}'"
-                )
+                log.output(f"Found HI-subtitle but skipping, 'cus hearing impaired is set to '{hi}'")
                 continue
             if hi is None and hi == "True":
-                log.output(
-                    f"Found nonHI-subtitle but skipping, 'cus hearing impaired is set to '{hi}'"
-                )
+                log.output(f"Found nonHI-subtitle but skipping, 'cus hearing impaired is set to '{hi}'")
                 continue
 
-            title_name = (
-                item.find("a", class_="bnone")
-                .text.replace("\n", "")
-                .replace("\t", "")
-                .replace("(", " (")
-            )
+            title_name = item.find("a", class_="bnone").text.replace("\n", "").replace("\t", "").replace("(", " (")
             log.output(f"{title_name} matched file hash")
-            th = [
-                a["href"]
-                for a in item.find_all(
-                    "a", href=lambda value: value and value.startswith("/en/subtitleserve/sub/")
-                )
-            ]
+            th = [a["href"] for a in item.find_all("a", href=lambda value: value and value.startswith("/en/subtitleserve/sub/"))]
             th = th[0] if th is not None else None
             link = f"https://www.opensubtitles.org{th}"
             download_url.append(link) if th is not None else None
