+++ SubSearch/src/util/string_parser.py
