+++ SubSearch/src/util/updates.py
@@ -1,13 +1,11 @@
 import cloudscraper
-from src.utilities import version
+from src.util import version
 
 SCRAPER = cloudscraper.create_scraper(browser={"browser": "chrome", "platform": "android", "desktop": False})
 
 
 def check_for_updates():
-    source = SCRAPER.get(
-        "https://raw.githubusercontent.com/vagabondHustler/SubSearch/main/SubSearch/src/data/version.json"
-    )
+    source = SCRAPER.get("https://raw.githubusercontent.com/vagabondHustler/SubSearch/main/SubSearch/src/data/version.json")
     scontent = source.content
     _string = str(scontent)
     _string_no_qoute = _string.replace('"', " ")
